LMS/requirements.txt
Django==4.2.2
djangorestframework==3.15.2
psycopg2==2.9.10
Pillow==11.0.0
django-filter==0.5.2